+++ src/backend/src/main/java/FoodHero/service/Dish/DishService.java
@@ -1,31 +1,89 @@
 package FoodHero.service.Dish;
 
 import FoodHero.dao.DishRepository;
+import FoodHero.model.Account;
 import FoodHero.model.Dish;
+import FoodHero.service.Account.AccountService;
+import FoodHero.service.Utils.ReturnCode;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
+
+import java.security.Principal;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Service
 public class DishService {
-    @Autowired
+
     private DishRepository dishRepository;
+    private AccountService accountService;
+    private static final Logger LOGGER = LogManager.getLogger(DishService.class);
 
-    public void createDish(Dish dish) {
+
+    @Autowired
+    public DishService(@Lazy DishRepository dishRepository, @Lazy AccountService accountService) {
+        this.dishRepository = dishRepository;
+        this.accountService = accountService;
+    }
+
+    public ReturnCode createDish(int id, Map<String, Object> payload) {
+        Account account = accountService.getAccount(id);
+        if (account == null) {
+            return ReturnCode.NOT_FOUND;
+        }
+        Dish dish = new Dish();
+        if (payload.get("name") != null && !payload.get("name").equals("") &&
+                payload.get("category") != null && !payload.get("category").equals("") &&
+                payload.get("description") != null && !payload.get("description").equals("")) {
+            dish.setAccount(account);
+            dish.setName((String) payload.get("name"));
+            dish.setCategory((String) payload.get("category"));
+            dish.setDescription((String) payload.get("description"));
+            dishRepository.save(dish);
+            return ReturnCode.OK;
+        }
+        return ReturnCode.INCORRECT_DATA;
+    }
+
+    public Dish getDish(int id) {
+        if (dishRepository.findById(id).isPresent()) {
+            return dishRepository.findById(id).get();
+        }
+        return null;
+    }
+
+    public ReturnCode updateDish(int id, Map<String, Object> payload) {
+        //TODO przy przepinaniu dań między użytkownikami trzeba zadbać o to aby tylko admin mogl to zrobić.
+        //jak narazie tego nei dodaje
+        Optional<Dish> dishOptional = dishRepository.findById(id);
+        if (!dishOptional.isPresent()) {
+            return ReturnCode.NOT_FOUND;
+        }
+        Dish dish = dishOptional.get();
+        if (payload.get("name") != null && !payload.get("name").equals("")) {
+            dish.setName((String) payload.get("name"));
+        }
+        if (payload.get("category") != null && !payload.get("category").equals("")) {
+            dish.setCategory((String) payload.get("category"));
+        }
+        if (payload.get("description") != null && !payload.get("description").equals("")) {
+            dish.setDescription((String) payload.get("description"));
+        }
         dishRepository.save(dish);
+        return ReturnCode.OK;
     }
 
-    public Optional<Dish> getDish(int id){
-        return Optional.ofNullable(dishRepository.getOne(id));
-    }
-
-    public void updateDish(Dish dish) {
-        dishRepository.save(dish);
-    }
-
-    public void deleteDish(int id) {
-        dishRepository.deleteById(id);
+    public ReturnCode deleteDish(int id) {
+        Optional<Dish> dish = dishRepository.findById(id);
+        if (dish.isPresent()) {
+            dishRepository.deleteById(id);
+            return ReturnCode.OK;
+        }
+        return ReturnCode.NOT_FOUND;
     }
 
     public List<Dish> getAllDishRaw() {
